+++ swap/tests/harness/mod.rs
@@ -867,7 +867,9 @@ impl Wallet for monero::Wallet {
     }
 
     async fn get_balance(&self) -> Result<Self::Amount> {
-        self.get_balance().await
+        let total = self.get_balance().await?;
+        let balance = Self::Amount::from_piconero(total.balance);
+        Ok(balance)
     }
 }